use hypercorn & add more error handling

+++ src/web_portal/views/admin.py
@@ -1,7 +1,10 @@
-from quart import Blueprint, redirect, render_template, request, url_for, flash
+import logging
+
+from quart import Blueprint, flash, redirect, render_template, request, url_for
+from tortoise.exceptions import DoesNotExist, IntegrityError
 
 from ..database import crud
-from ..helpers import login_admin_required
+from ..helpers import PasswordStrength, login_admin_required
 
 blueprint = Blueprint("admin", __name__, url_prefix="/admin")
 
@@ -20,12 +23,19 @@ async def index():
 @blueprint.route("/new-widget", methods=["POST"])
 @login_admin_required
 async def new_widget():
-    url = (await request.form)["url"]
-    group_id = (await request.form)["group_id"]
-    prefix = (await request.form)["prefix"]
-    color_name = (await request.form)["color_name"]
-    await crud.new_panel_widget(url, prefix, color_name, group_id)
-    await flash("created new widget", "green")
+    try:
+        url = (await request.form)["url"]
+        group_id = (await request.form)["group_id"]
+        if not group_id.isnumeric():
+            await flash("form field type error", "red")
+        else:
+            prefix = (await request.form)["prefix"]
+            color_name = (await request.form)["color_name"]
+            await crud.new_panel_widget(url, prefix, color_name, group_id)
+            await flash("created new widget", "green")
+    except IntegrityError:
+        logging.exception("new_widget IntegrityError")
+        await flash("form field id's are invalid", "red")
     return redirect(url_for("admin.index"))
 
 @blueprint.route("/new-group", methods=["POST"])
@@ -39,18 +49,28 @@ async def new_group():
 @blueprint.route("/re-goup-widget", methods=["POST"])
 @login_admin_required
 async def re_group_widget():
-    widget_id = (await request.form)["widget_id"]
-    group_id = (await request.form)["group_id"]
-    await crud.modify_widget_group(widget_id, group_id)
-    await flash("changed widget group", "green")
+    try:
+        widget_id = (await request.form)["widget_id"]
+        group_id = (await request.form)["group_id"]
+        if not widget_id.isnumeric() or not group_id.isnumeric():
+            await flash("form field type error", "red")
+        else:
+            await crud.modify_widget_group(widget_id, group_id)
+            await flash("changed widget group", "green")
+    except IntegrityError:
+        logging.exception("new_widget IntegrityError")
+        await flash("form field id's are invalid", "red")
     return redirect(url_for("admin.index"))
 
 @blueprint.route("/delete-widget", methods=["POST"])
 @login_admin_required
 async def delete_widget():
     widget_id = (await request.form)["widget_id"]
-    await crud.delete_widget_by_id(widget_id)
-    await flash("deleted widget", "green")
+    if not widget_id.isnumeric():
+        await flash("form field type error", "red")
+    else:
+        await crud.delete_widget_by_id(widget_id)
+        await flash("deleted widget", "green")
     return redirect(url_for("admin.index"))
 
 @blueprint.route("/change-widget-color", methods=["POST"])
@@ -58,26 +78,37 @@ async def delete_widget():
 async def change_widget_color():
     widget_id = (await request.form)["widget_id"]
     color_name = (await request.form)["color_name"]
-    await crud.modify_widget_color(widget_id, color_name)
-    await flash("changed widget color", "green")
+    if not widget_id.isnumeric():
+        await flash("form field type error", "red")
+    else:
+        await crud.modify_widget_color(widget_id, color_name)
+        await flash("changed widget color", "green")
     return redirect(url_for("admin.index"))
 
 @blueprint.route("/new-user", methods=["POST"])
 @login_admin_required
 async def new_user():
-    username = (await request.form)["username"]
-    password = (await request.form)["password"]
-    is_admin = (await request.form).get("is_admin", False, bool)
-    await crud.new_user(username, password, is_admin)
-    await flash("created new user", "green")
+    try:
+        username = (await request.form)["username"]
+        password = (await request.form)["password"]
+        if len(password) < 8:
+            raise PasswordStrength("password less than 8 characters")
+        is_admin = (await request.form).get("is_admin", False, bool)
+        await crud.new_user(username, password, is_admin)
+        await flash("created new user", "green")
+    except PasswordStrength as err:
+        await flash(err.args[0], "red")
     return redirect(url_for("admin.index"))
 
 @blueprint.route("/delete-user", methods=["POST"])
 @login_admin_required
 async def delete_user():
     user_id = (await request.form)["user_id"]
-    await crud.delete_user(user_id)
-    await flash("deleted user", "green")
+    if not user_id.isnumeric():
+        await flash("form field type error", "red")
+    else:
+        await crud.delete_user(user_id)
+        await flash("deleted user", "green")
     return redirect(url_for("admin.index"))
 
 @blueprint.route("/modify-user-permissions", methods=["POST"])
@@ -85,15 +116,29 @@ async def delete_user():
 async def modify_user_permissions():
     user_id = (await request.form)["user_id"]
     is_admin = (await request.form).get("is_admin", False, bool)
-    await crud.modify_user_permissions(user_id, is_admin)
-    await flash("modified permissions", "green")
+    if not user_id.isnumeric():
+        await flash("form field type error", "red")
+    else:
+        await crud.modify_user_permissions(user_id, is_admin)
+        await flash("modified permissions", "green")
     return redirect(url_for("admin.index"))
 
 @blueprint.route("/change-password", methods=["POST"])
 @login_admin_required
 async def change_user_password():
-    user_id = (await request.form)["user_id"]
-    new_password = (await request.form)["new_password"]
-    await crud.modify_user_password(user_id, new_password)
-    await flash("changed password", "green")
+    try:
+        user_id = (await request.form)["user_id"]
+        new_password = (await request.form)["new_password"]
+        if len(new_password) < 8:
+            raise PasswordStrength("password less than 8 characters")
+        if not user_id.isnumeric():
+            await flash("form field type error", "red")
+        else:
+            await crud.modify_user_password(user_id, new_password)
+            await flash("changed password", "green")
+    except DoesNotExist:
+        logging.exception("row with id not found")
+        await flash("user does not exist with given id", "red")
+    except PasswordStrength as err:
+        await flash(err.args[0], "red")
     return redirect(url_for("admin.index"))
